Consolidate node module rules & providers & support transitive npm deps via collect_node_modules_aspect
