--- internal/ng_apf_library/BUILD.bazel
@@ -1,40 +1,0 @@
-# Copyright 2019 The Bazel Authors. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-load("@bazel_skylib//:bzl_library.bzl", "bzl_library")
-
-licenses(["notice"])  # Apache 2.0
-
-package(default_visibility = ["//internal:__subpackages__"])
-
-bzl_library(
-    name = "bzl",
-    srcs = glob(["*.bzl"]),
-    visibility = ["//visibility:public"],
-)
-
-filegroup(
-    name = "package_contents",
-    srcs = [
-        "BUILD.bazel",
-        "ng_apf_library.bzl",
-        "ng_apf_library.js",
-    ],
-    visibility = ["//:__pkg__"],
-)
-
-exports_files([
-    "ng_apf_library.bzl",
-    "ng_apf_library.js",
-])
