+++ internal/npm_install/npm_install.bzl
@@ -107,12 +107,6 @@ def _add_scripts(repository_ctx):
         {},
     )
 
-    repository_ctx.template(
-        "ng_apf_library.js",
-        repository_ctx.path(Label("//internal/ng_apf_library:ng_apf_library.js")),
-        {},
-    )
-
 def _add_data_dependencies(repository_ctx):
     """Add data dependencies to the repository."""
     for f in repository_ctx.attr.data:
