+++ internal/npm_install/test/golden/BUILD.bazel.golden
@@ -1,6 +1,7 @@
 
 package(default_visibility = ["//visibility:public"])
-filegroup(
+load("@build_bazel_rules_nodejs//internal/npm_install:node_module_library.bzl", "node_module_library")
+node_module_library(
     name = "node_modules",
     srcs = [
         "//node_modules/.bin:.bin__files",
